Add a check on the returned round value
The beacon server might not returned the round the user asked for. If enabled, the verification should fail in this case.
This commit also guards `HttpClient::latest` behinds the `time` feature.
If `time` is not available, the method `HttpClient.get` has to be called instead.

The commit is co-authored, based on 8448548.

Co-authored-by: Patrick McClurg <patrick.mcclurg@protocol.ai>
Co-authored-by: Thibault <contact@thibaultmeunier.com>